requirements.txt
asgiref==3.8.1
Django==4.2.16
djangorestframework==3.15.2
dnspython==2.7.0
MarkupSafe==3.0.2
mongoengine==0.29.1
pymongo==4.10.1
python-dotenv==1.0.1
sqlparse==0.5.1
typing_extensions==4.12.2
tzdata==2024.2
Werkzeug==3.1.2
